// src/generated/types/LabListDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { UserDto } from './UserDto.ts'

export const labListDtoStatusEnum = {
  Available: 'Available',
  'In Use': 'In Use',
  'Under Maintenance': 'Under Maintenance',
} as const

export type LabListDtoStatusEnum = (typeof labListDtoStatusEnum)[keyof typeof labListDtoStatusEnum]

export type LabListDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  location: string
  /**
   * @type string
   */
  supervisorId: string
  /**
   * @type string
   */
  id: string
  /**
   * @type string, date-time
   */
  createdAt: Date
  /**
   * @type string, date-time
   */
  updatedAt: Date
  /**
   * @type number
   */
  deviceCount: number
  /**
   * @type object
   */
  supervisor: UserDto
  /**
   * @type string
   */
  status: LabListDtoStatusEnum
}