// src/generated/types/UpdateLabDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type UpdateLabDto = {
  /**
   * @type string | undefined
   */
  name?: string | undefined
  /**
   * @type string | undefined
   */
  location?: string | undefined
  /**
   * @type string | undefined
   */
  supervisorId?: string | undefined
}